1. Full management commands system upgrade. Commands system now uses `click` library which give more flexiblity and is much better than initial/manual implementation. 2. Update default commands to use new commands(click) syntax. 3. Added new async to sync function api
